# Speed up `UnitBase.compose()`