--- a/docs/changes/units/17425.perf.rst
+++ b/docs/changes/units/17425.perf.rst
@@ -1,0 +1,1 @@
+``UnitBase.compose()`` is now 20% faster.