### Description

I've simplified the function and managed to speed it up by 20% or so in the process. I've opened this pull request with two commits because I'm not sure if the simplifications in the second commit are allowed. If they are allowed then the commits should be combined into one before merging. If they are not allowed then the second commit should just be dropped.

### Performance comparison

On `main`:
```python
Python 3.12.3 (main, Nov  6 2024, 18:32:19) [GCC 13.2.0]
Type 'copyright', 'credits' or 'license' for more information
IPython 8.27.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: from astropy import units as u

In [2]: %timeit u.kg.compose(units=u.si)
817 μs ± 2.1 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [3]: %timeit u.Pa.compose(units=u.si)
1.59 ms ± 5.72 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```
With the patch here:
```python
In [1]: from astropy import units as u

In [2]: %timeit u.kg.compose(units=u.si)
647 μs ± 944 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [3]: %timeit u.Pa.compose(units=u.si)
1.21 ms ± 1.52 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

- [ ] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.