@nstarman 2024-11-21
The first commit looks great. Haven't looked at 2nd yet.